requirements.txt
botcity-framework-web>=0.8.0
botcity-maestro-sdk>=0.3.3
botcity-email-plugin
xlsxwriter